Fix for a bug in firewalld changing defaultzone
This is a fix for a bug in firewalld when you change the DefaultZone at the same time as adding the new zone that will be the DefaultZone without doing a `firewall-cmd --reload` inbetween so that firewalld would know about the new zone ahead of time.  This adds that inbetween step until the bug gets recognized and fixed.  This is an elegant solution I believe as it doesn't allow for firewalld to even hit the bug, which it would if we just did the changes and ran `firewall-cmd --reload; firewall-cmd --reload' which could potentially cause a short firewall drop.